+++ cdk.json
@@ -70,6 +70,26 @@
     "@aws-cdk/aws-ec2:ebsDefaultGp3Volume": true,
     "@aws-cdk/aws-ecs:removeDefaultDeploymentAlarm": true,
     "@aws-cdk/custom-resources:logApiResponseDataPropertyTrueDefault": false,
-    "@aws-cdk/aws-s3:keepNotificationInImportedBucket": false
+    "@aws-cdk/aws-s3:keepNotificationInImportedBucket": false,
+    "stackName": "undefined",
+    "VERSION_NUMBER": "undefined",
+    "COMMIT_ID": "undefined",
+    "useMockOidc": "undefined",
+    "primaryOidcClientId": "undefined",
+    "primaryOidClientSecret": "undefined",
+    "primaryOidcIssuer": "undefined",
+    "primaryOidcAuthorizeEndpoint": "undefined",
+    "primaryOidcTokenEndpoint": "undefined",
+    "primaryOidcUserInfoEndpoint": "undefined",
+    "primaryOidcjwksEndpoint": "undefined",
+    "mockOidcClientId": "undefined",
+    "mockOidClientSecret": "undefined",
+    "mockOidcIssuer": "undefined",
+    "mockOidcAuthorizeEndpoint": "undefined",
+    "mockOidcTokenEndpoint": "undefined",
+    "mockOidcUserInfoEndpoint": "undefined",
+    "mockOidcjwksEndpoint": "undefined",
+    "epsDomain": "undefined",
+    "epsZoneId": "undefined"
   }
 }
